refactor: rename ui package to output
