+++ internal/output/progress.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 Rich Haase
 // Licensed under the MIT License. See LICENSE file in the project root for license information.
 
-package ui
+package output
 
 import (
 	"fmt"
