+++ internal/output/tables.go
@@ -1,9 +1,9 @@
 // Copyright (c) 2025 Rich Haase
 // Licensed under the MIT License. See LICENSE file in the project root for license information.
 
-// Package ui contains all user interface rendering logic including
+// Package output contains all user interface rendering logic including
 // tables, formatters, and output presentation.
-package ui
+package output
 
 // This file will contain table rendering logic
 // currently in internal/commands/shared.go
